backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.0
pytest==7.4.3
pytest-asyncio==0.23.2

